renamed to detect.py
